+++ test/widgets/board_test.dart
@@ -2168,7 +2168,7 @@ void main() {
     late ChessboardController controller;
 
     setUp(() {
-      controller = ChessboardController(initialFen: kInitialFEN);
+      controller = ChessboardController.nonInteractive(initialFen: kInitialFEN);
     });
 
     tearDown(() {
@@ -2263,7 +2263,7 @@ void main() {
     late ChessboardController controller;
 
     setUp(() {
-      controller = ChessboardController(initialFen: kInitialFEN);
+      controller = ChessboardController.nonInteractive(initialFen: kInitialFEN);
     });
 
     tearDown(() {
@@ -2338,7 +2338,7 @@ void main() {
       await tester.pump();
 
       // Controller should still drive the board correctly.
-      controller.jumpToPosition('rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1');
+      controller.jumpToFen('rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1');
       await tester.pump();
 
       expect(_piecesPainter(tester).pieces.containsKey(Square.e4), isTrue);
@@ -2355,15 +2355,12 @@ void main() {
       bool showExtra = false;
       const position = Chess.initial;
       final interactiveController = ChessboardController(
-        initialFen: kInitialFEN,
         initialGame: GameData(
+          fen: kInitialFEN,
           playerSide: PlayerSide.both,
           isCheck: false,
           sideToMove: Side.white,
           validMoves: makeLegalMoves(position),
-          promotionMove: null,
-          onMove: (_, {viaDragAndDrop}) {},
-          onPromotionSelection: (_) {},
         ),
       );
       addTearDown(interactiveController.dispose);
@@ -2379,6 +2376,8 @@ void main() {
                     controller: interactiveController,
                     size: boardSize,
                     orientation: Side.white,
+                    onMove: (_, {viaDragAndDrop}) {},
+                    onPromotionSelection: (_) {},
                   ),
                   ElevatedButton(
                     onPressed: () => setState(() => showExtra = !showExtra),
@@ -2414,15 +2413,12 @@ void main() {
       bool showExtra = true;
       const position = Chess.initial;
       final interactiveController = ChessboardController(
-        initialFen: kInitialFEN,
         initialGame: GameData(
+          fen: kInitialFEN,
           playerSide: PlayerSide.both,
           isCheck: false,
           sideToMove: Side.white,
           validMoves: makeLegalMoves(position),
-          promotionMove: null,
-          onMove: (_, {viaDragAndDrop}) {},
-          onPromotionSelection: (_) {},
         ),
       );
       addTearDown(interactiveController.dispose);
@@ -2438,6 +2434,8 @@ void main() {
                     controller: interactiveController,
                     size: boardSize,
                     orientation: Side.white,
+                    onMove: (_, {viaDragAndDrop}) {},
+                    onPromotionSelection: (_) {},
                   ),
                   ElevatedButton(
                     onPressed: () => setState(() => showExtra = !showExtra),
@@ -2490,7 +2488,7 @@ void main() {
       for (var i = 0; i < 3; i++) {
         await tester.tap(find.text('toggle'));
         await tester.pump();
-        controller.jumpToPosition(i.isEven ? altFen : kInitialFEN);
+        controller.jumpToFen(i.isEven ? altFen : kInitialFEN);
         await tester.pump();
       }
 
@@ -2661,12 +2659,9 @@ class _TestAppState extends State<_TestApp> {
     interactiveSide = widget.initialPlayerSide;
     position = Position.setupPosition(widget.rule, Setup.parseFen(widget.initialFen));
     promotionMove = widget.initialPromotionMove;
+    lastMove = widget.initialPromotionMove;
     shapes = widget.initialShapes ?? {};
-    _controller = ChessboardController(
-      initialFen: position.fen,
-      initialGame: _buildGame(),
-      initialLastMove: widget.initialPromotionMove,
-    );
+    _controller = ChessboardController(initialGame: _buildGame());
     _gameEventSub = widget.gameEventStream?.listen(_onGameEvent);
   }
 
@@ -2679,27 +2674,15 @@ class _TestAppState extends State<_TestApp> {
 
   GameData _buildGame() {
     return GameData(
+      fen: position.fen,
+      lastMove: lastMove,
       playerSide: interactiveSide,
       isCheck: position.isCheck,
       sideToMove: position.turn == Side.white ? Side.white : Side.black,
       validMoves: makeLegalMoves(position),
       promotionMove: promotionMove,
       droppable: widget.droppable,
-      onMove: _onMove,
-      onPromotionSelection: (Role? role) {
-        if (role != null) {
-          _playMove(promotionMove!.withPromotion(role));
-        }
-        promotionMove = null;
-        _controller.updatePosition(position.fen, game: _buildGame(), lastMove: lastMove);
-      },
-      premovable: (
-        premove: premoveData,
-        onSetPremove: (Move? move) {
-          premoveData = move;
-          _controller.updatePosition(position.fen, game: _buildGame());
-        },
-      ),
+      premovable: (premove: premoveData),
       canPromoteToKing: widget.canPromoteToKing,
     );
   }
@@ -2708,10 +2691,10 @@ class _TestAppState extends State<_TestApp> {
     switch (event) {
       case GameEvent.nonInteractiveBoardEvent:
         interactiveSide = PlayerSide.none;
-        _controller.updatePosition(position.fen, game: _buildGame());
+        _controller.updatePosition(_buildGame());
       case GameEvent.interactiveBoardEvent:
         interactiveSide = widget.initialPlayerSide;
-        _controller.updatePosition(position.fen, game: _buildGame());
+        _controller.updatePosition(_buildGame());
       case GameEvent.externalMove:
         final allMoves = [
           for (final entry in position.legalMoves.entries)
@@ -2720,7 +2703,7 @@ class _TestAppState extends State<_TestApp> {
         if (allMoves.isNotEmpty) {
           position = position.playUnchecked(allMoves.first);
           lastMove = allMoves.first;
-          _controller.updatePosition(position.fen, game: _buildGame(), lastMove: lastMove);
+          _controller.updatePosition(_buildGame());
         }
     }
   }
@@ -2747,12 +2730,7 @@ class _TestAppState extends State<_TestApp> {
     } else {
       _playMove(move);
     }
-    _controller.updatePosition(
-      position.fen,
-      game: _buildGame(),
-      lastMove: lastMove,
-      lastDrop: viaDragAndDrop == true ? move : null,
-    );
+    _controller.updatePosition(_buildGame(), lastDrop: viaDragAndDrop == true ? move : null);
 
     if (widget.shouldPlayOpponentMove) {
       Timer(const Duration(milliseconds: 200), () {
@@ -2766,7 +2744,7 @@ class _TestAppState extends State<_TestApp> {
           interactiveSide = PlayerSide.none;
         }
         lastMove = opponentMove;
-        _controller.updatePosition(position.fen, game: _buildGame(), lastMove: opponentMove);
+        _controller.updatePosition(_buildGame());
 
         // play premove just after the opponent move
         if (premoveData != null) {
@@ -2775,13 +2753,13 @@ class _TestAppState extends State<_TestApp> {
               scheduleMicrotask(() {
                 position = position.playUnchecked(premoveData!);
                 premoveData = null;
-                _controller.updatePosition(position.fen, game: _buildGame(), lastMove: lastMove);
+                _controller.updatePosition(_buildGame());
               });
             } else {
               scheduleMicrotask(() {
                 promotionMove = premoveData as NormalMove?;
                 premoveData = null;
-                _controller.updatePosition(position.fen, game: _buildGame());
+                _controller.updatePosition(_buildGame());
               });
             }
           }
@@ -2802,6 +2780,18 @@ class _TestAppState extends State<_TestApp> {
               size: boardSize,
               settings: widget.settings ?? defaultSettings,
               orientation: widget.orientation,
+              onMove: _onMove,
+              onPromotionSelection: (Role? role) {
+                if (role != null) {
+                  _playMove(promotionMove!.withPromotion(role));
+                }
+                promotionMove = null;
+                _controller.updatePosition(_buildGame());
+              },
+              onSetPremove: (Move? move) {
+                premoveData = move;
+                _controller.updatePosition(_buildGame());
+              },
               onTouchedSquare: widget.onTouchedSquare,
               shapes: shapes,
             ),